+++ src/main/java/org/pengu/model/CakeValueOverWeight.java
@@ -4,6 +4,6 @@ import java.util.Comparator;
 
 public class CakeValueOverWeight implements Comparator<Cake> {
     public int compare(Cake cakeA, Cake cakeB) {
-        return cakeB.getValueOverWeight() - cakeA.getValueOverWeight();
+        return (int) (cakeB.getValueOverWeight() - cakeA.getValueOverWeight());
     }
 }
